Moved io functions from console to parsing
