+++ sarada/console.py
@@ -8,7 +8,7 @@ import os
 import sys
 
 from pathlib import Path
-from typing import Final, Iterable, Iterator, List
+from typing import Final, List
 
 import typer
 
@@ -16,13 +16,12 @@ from loguru import logger
 
 from sarada.logging import setup_logging
 from sarada.neuron import Neuron
-from sarada.notebook import Notebook, Pitch
+from sarada.notebook import Pitch
 from sarada.numeris import Numeris
-from sarada.parsing import create_stream, extract_notes
+from sarada.parsing import read_scores, store_score
 
-app = typer.Typer()
+app: Final = typer.Typer()
 
-supported_extensions: Final = [".abc"]
 window_size: Final = 100
 
 arg_music_dir = typer.Argument(..., help="Path to directory containing learnign data")
@@ -113,36 +112,7 @@ def generate(model_path: Path = arg_model_path) -> None:
     sequence = model.generate(100)
     pitches = numeris.denumerize(sequence)
 
-    store_sequence(pitches)
-
-
-def read_scores(path: Path, recursive: bool = False) -> Notebook:
-    """
-    Open file on given path and aggregate them in Notebook instance.
-    """
-    logger.info("Processing files in {path}", path=str(path))
-
-    scores = read_files(path, recursive)
-    notes_source = extract_notes(scores)
-
-    notes = Notebook()
-    for note in notes_source:
-        notes.add(note)
-
-    logger.info("Finished loading files")
-
-    return notes
-
-
-def store_sequence(pitches: Iterable[Pitch]) -> None:
-    """
-    Store sequence in midi file.
-    """
-    path = "test_output.mid"
-    logger.info("Storing sequence at {path}", path=path)
-    stream = create_stream(pitches)
-    logger.debug("Saving file in {path}", path=path)
-    stream.write("midi", fp=path)
+    store_score(pitches)
 
 
 def load_data(model_path: Path) -> Numeris[Pitch]:
@@ -158,24 +128,6 @@ def load_data(model_path: Path) -> Numeris[Pitch]:
     numeris = Numeris(data)
 
     return numeris
-
-
-def read_files(path: Path, recursive: bool) -> Iterator[str]:
-    """
-    Iterate over content of musical files in provided directory.
-    """
-    for filepath in path.iterdir():
-        if filepath.is_file() and filepath.suffix in supported_extensions:
-            logger.debug("Opening file {path}", path=filepath)
-            with open(filepath, encoding="utf-8") as audio_file:
-                yield audio_file.read()
-        elif recursive and filepath.is_dir():
-            logger.debug("Searching {path}", path=filepath)
-            yield from read_files(filepath, recursive=True)
-        else:
-            logger.debug(
-                "File {name} omitted due to unsupported extension", name=filepath
-            )
 
 
 if __name__ == "__main__":
