+++ sarada/parsing.py
@@ -3,12 +3,16 @@ Parse raw file content into workable streams.
 """
 from __future__ import annotations
 
-from typing import Iterable, Iterator, List, Optional
+from pathlib import Path
+from typing import Final, Iterable, Iterator, List, Optional
 
+from loguru import logger
 from music21 import converter, instrument
 
 from sarada import music21
-from sarada.notebook import Pitch
+from sarada.notebook import Notebook, Pitch
+
+supported_extensions: Final = [".abc"]
 
 
 def extract_notes(
@@ -43,3 +47,50 @@ def create_stream(pitches: Iterable[Pitch]) -> music21.Stream:
         notes.append(note)
 
     return music21.Stream(notes)
+
+
+def read_scores(path: Path, recursive: bool = False) -> Notebook:
+    """
+    Open file on given path and aggregate them in Notebook instance.
+    """
+    logger.info("Processing files in {path}", path=str(path))
+
+    scores = read_files(path, recursive)
+    notes_source = extract_notes(scores)
+
+    notes = Notebook()
+    for note in notes_source:
+        notes.add(note)
+
+    logger.info("Finished loading files")
+
+    return notes
+
+
+def read_files(path: Path, recursive: bool) -> Iterator[str]:
+    """
+    Iterate over content of musical files in provided directory.
+    """
+    for filepath in path.iterdir():
+        if filepath.is_file() and filepath.suffix in supported_extensions:
+            logger.debug("Opening file {path}", path=filepath)
+            with open(filepath, encoding="utf-8") as audio_file:
+                yield audio_file.read()
+        elif recursive and filepath.is_dir():
+            logger.debug("Searching {path}", path=filepath)
+            yield from read_files(filepath, recursive=True)
+        else:
+            logger.debug(
+                "File {name} omitted due to unsupported extension", name=filepath
+            )
+
+
+def store_score(pitches: Iterable[Pitch]) -> None:
+    """
+    Store sequence in midi file.
+    """
+    path = "test_output.mid"
+    logger.info("Storing sequence at {path}", path=path)
+    stream = create_stream(pitches)
+    logger.debug("Saving file in {path}", path=path)
+    stream.write("midi", fp=path)
